реализовал перевод между своими клиентами, пока без валидации и толковых сообщений

+++ src/main/java/net/snopser/bank/snopserbank/repository/AccountRepository.java
@@ -3,9 +3,11 @@ package net.snopser.bank.snopserbank.repository;
 import net.snopser.bank.snopserbank.entity.Account;
 import org.springframework.data.jpa.repository.JpaRepository;
 
+import java.math.BigInteger;
+
 /**
  * @author Виктор Фалькенберг (viktor.falkenberg@mediascope.net)
  */
 
-public interface AccountRepository extends JpaRepository<Account, Integer> {
+public interface AccountRepository extends JpaRepository<Account, BigInteger> {
 }
